Fix loading for more parsers
