--- dojo/tools/crashtest_security_xml/__init__.py
@@ -1,1 +1,0 @@
-__author__ = 'phylu'
