+++ dojo/tools/nexpose/parser.py
@@ -5,7 +5,7 @@ from defusedxml import ElementTree
 from dojo.models import Finding, Endpoint
 
 
-class NexposeFullXmlParser(object):
+class NexposeParser(object):
     """
     The objective of this class is to parse Nexpose's XML 2.0 Report.
 
@@ -15,6 +15,15 @@ class NexposeFullXmlParser(object):
 
     @param xml_filepath A proper xml generated by nexpose
     """
+
+    def get_scan_types(self):
+        return ["Nexpose Scan"]
+
+    def get_label_for_scan_types(self, scan_type):
+        return scan_type  # no custom label for now
+
+    def get_description_for_scan_types(self, scan_type):
+        return "Use the full XML export template from Nexpose."
 
     def get_findings(self, xml_output, test):
         tree = ElementTree.parse(xml_output)
