+++ dojo/tools/ossindex_devaudit/parser.py
@@ -9,6 +9,16 @@ class OssIndexDevauditParser(object):
     Parses files created by the Sonatype OssIndex Devaudit tool
     https://github.com/sonatype-nexus-community/DevAudit
     """
+
+    def get_scan_types(self):
+        return ["OssIndex Devaudit SCA Scan Importer"]
+
+    def get_label_for_scan_types(self, scan_type):
+        return scan_type  # no custom label for now
+
+    def get_description_for_scan_types(self, scan_type):
+        return "Import OssIndex Devaudit SCA Scan in json format."
+
     def get_findings(self, json_file, test):
 
         tree = self.parse_json(json_file)
