+++ dojo/unittests/tools/test_crashtest_security_parser.py
@@ -1,19 +1,19 @@
 from django.test import TestCase
-from dojo.tools.crashtest_security_json.parser import CrashtestSecurityJsonParser
+from dojo.tools.crashtest_security.parser import CrashtestSecurityParser
 from dojo.models import Test
 
 
-class TestCrashtestSecurityJsonParser(TestCase):
+class TestCrashtestSecurityParser(TestCase):
     def test_crashtest_security_json_parser_empty_file_has_no_findings(self):
         testfile = open("dojo/unittests/scans/crashtest_security_json/empty.json")
-        parser = CrashtestSecurityJsonParser()
+        parser = CrashtestSecurityParser()
         findings = parser.get_findings(testfile, Test())
         testfile.close()
         self.assertEqual(0, len(findings))
 
     def test_crashtest_security_json_parser_full_file_has_many_findings(self):
         testfile = open("dojo/unittests/scans/crashtest_security_json/full.json")
-        parser = CrashtestSecurityJsonParser()
+        parser = CrashtestSecurityParser()
         findings = parser.get_findings(testfile, Test())
         testfile.close()
         self.assertEqual(73, len(findings))
@@ -22,7 +22,7 @@ class TestCrashtestSecurityJsonParser(TestCase):
         testfile = open(
             "dojo/unittests/scans/crashtest_security_json/data_extracted.json"
         )
-        parser = CrashtestSecurityJsonParser()
+        parser = CrashtestSecurityParser()
         findings = parser.get_findings(testfile, Test())
         testfile.close()
         self.assertEqual(73, len(findings))
